18662: Fixed - 'Disassembler' has not been declared
