+++ hotspot/src/cpu/mips/vm/nativeInst_mips.cpp
@@ -25,6 +25,7 @@
 
 #include "precompiled.hpp"
 #include "asm/macroAssembler.hpp"
+#include "code/codeBlob.hpp"
 #include "code/codeCache.hpp"
 #include "compiler/disassembler.hpp"
 #include "memory/resourceArea.hpp"
